adds a marquee selection & editing tool
* adds a ToolSelector on the left with two tools, arrow & marquee

* arrow Tool functions like the old UI

* marquee allows you to click and drag a rectangular marquee and then click and drag selected hexagons

* dragging multi selections works pretty much the same as a singular one

* Known Issue: dragging new hexagons onto the canvas doesn't work in marquee mode

The big question is whether or not the arrow tool is even still useful. Right now I've left a bit of GridGraphic.js a little messy. There could be a lot of combination of the arrow selection method and the marquee selection method. If we would like to keep both tools I'll probably refactor some of the arrow selection stuff to be more like the marquee selection.